Update add checkpoint to use long instead of int for file id.

+++ servers/src/main/java/tachyon/worker/block/BlockDataManager.java
@@ -143,7 +143,7 @@ public final class BlockDataManager {
    * @throws TException if the file does not exist or cannot be renamed
    * @throws IOException if the update to the master fails
    */
-  public void addCheckpoint(long userId, int fileId) throws TException, IOException {
+  public void addCheckpoint(long userId, long fileId) throws TException, IOException {
     // TODO This part needs to be changed.
     String srcPath = PathUtils.concatPath(getUserUfsTmpFolder(userId), fileId);
     String ufsDataFolder =
